Changed the error checking to detect the element types supported.
Changed the order of events slightly when starting an animation to make 
them more sensible.

Fixed a bug where it tried to create a sprite when one wasn't returned 
from spriteDatas.

Slowed down the speed of the scroll animation to be closer to the 
triangle scroll speed.

Fixed the graphical glitches during the triangle scroll animation.

Added the text element type which is almost identical in syntax to the 
sprite type.

Sprite datas can now be returned to be created in 
animationType.menuSprite.init.

Updated Bagel.js

Added text to the demo.