Added scrolling and a circle animation.
The "circle" animation is currently unfinished as it doesn't have a 
reverse yet.
It takes these arguments:
x, y, color and initialSize.

Buttons now make some arguments for animations optional. If unspecified 
and a valid argument for the animation, x, y, size and initialSize (uses 
element.size) will be set to the corresponding values.

The camera's position is now reset when changing submenus.

Hopefully fixed a potential softlock in the scroll animation (should 
probably test it).

Updated the scroll animation description to include the new subtypes.

Added a check function in the animation JSON. Moved the animation checks 
for the button into its JSON.

menuSprite.animateSubmenuChange accepts 2 more optional arguments as 
part of the button animation changes.

The button onclick function is now called with the sprite and menuSprite 
as the second and third arguments.

Updated Bagel.js to use the new scrolling input.