Some tweaks, bug fixes and updated Bagel.js.
Removed the scroll speed option as it's now part of Bagel.js.

Added menuSprite.submenus.<submenu>.init which runs once the animation 
has finished.

Added fixedToCamera as an option for elements.

Removed the coordinate sprite and fixed anchors, mostly by updating 
Bagel.js.

Added getFutureID as the 3rd argument to elementTypeJSON.spriteDatas. It 
returns the id that the sprite that many times ahead will have. Useful 
for linking sprites together.

Added minProcess and deleteOnOtherDelete as options when creating a 
sprite in spriteDatas.

Sprites returned from spriteDatas no longer need a stateToRun for their 
scripts as it's set to game.state if unspecified (no reason for it to be 
anything else).

Added elementTypeJSON.preload which runs when the menu is first 
initialised, allowing assets to be loaded in a loading screen. It still 
needs to load based on the submenu.