Moved where elements are created, added the circle close animation and fixed some bugs
Elements are now stored in menuSprite.submenus.<>.elements. 

Removed the submenu argument for elements.

Fixed a bug where you could scroll during animations?

The camera's position is now saved between submenus and is restored to 
the position it was when that submenu was last visited.

Added the close direction for the circle animation.