Added hover text, some new internal methods and renamed onclick to onClick.
Set the hover text for a button by setting onHover to the text. Or set 
it to a function that's called with the same arguments as onClick.

The hover text can be changed in menuSprite.submenus.<>.hoverText.

Added the internal methods queueMakeSpriteElement, makeSpriteElement and 
addSpriteElement. Queue runs makeSpriteElement near the end of the frame 
so it will be on top. makeSpriteElement takes the sprite data and makes 
it into a spriteElement (it's returned as well as added to 
menuSprite.internal.spriteElements). And addSpriteElement takes an 
already existing sprite and adds it to spriteElements so it's deleted on 
submenu changes and menuSprite deletions.

Fixed a bug which caused anchors to behave incorrectly for elements.

Button icons now keep their original aspect ratio.

Updated Bagel.js.