Reduced the resolution of the circle animation to improve performance.
The circle close animation resolution is now based around the game size 
rather than the size it will show at. This should help with performance 
while it still looks fairly good.

Added a resolution option to the circle animation. Defaults to 0.5 / the 
device pixel ratio. Slightly lower than the original which was 0.8 (and 
it wasn't divided by the ratio). Still looks comparable though.