Updated Bagel.js, made elements slightly more modualar and added fading to hover text.
Every element type now has its own variables. Default values can be set 
in the initType method which is called with elementTypeVars, the game 
and if it's a rerun or not (it's run every time the submenu is 
initialised but also when the menuSprite is first made)

hoverTextSprite and hoverText is now stored in this instead of directly 
in menuSprite.internal. However, it's still not completely modular as 
the syntax for the hover text options has to be hard coded in at the 
moment.